+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * Using NSNumber properties (unsupported) now throws a more informative exception.
 * Added `[RLMRealm defaultRealmPath]`
 * Proper implementation for [RLMArray indexOfObjectWhere:]
+* The default Realm path on OS X is now ~/Library/Application Support/[bundle
+  identifier]/default.realm rather than ~/Documents
 
 ### Bugfixes
 
@@ -19,6 +21,8 @@
 * Fixed potential bug related to the handling of array properties (RLMArray).
 * Fixed bug where array properties accessed the wrong property.
 * Fixed bug that prevented objects with custom getters to be added to a Realm.
+* Fixed a bug where initializing a standalone object with an array literal would 
+  trigger an exception.
 
 
 0.81.0 Release notes (2014-07-22)